One page shell; Inbox absorbs Messaging routing (§28)
Automations, Activity, and Inbox each had their own indentation and head
style, and inbox-delivery config hid under Connectors with the mirror
channel editable in two places.

- Every top-level page now uses the Connectors/Activity shell: centered
  column, PanelHead, cards. Automations drops the icon-title and the
  boxed banner; its list/detail/templates restyle onto the shell.
- Inbox gains page-level tabs: Pending (the queue, unchanged) and
  Configure — the former Messaging-routing page moved whole into
  InboxConfigure.tsx (mirror, DM route, subscriptions, Unrouted with
  the warning badge). Pending's routing line is read-only and links to
  Configure; the duplicate inline mirror editor is deleted. The mirror
  target and routing line show channel names when known.
- Connectors sub-nav shrinks to Connectors and MCP servers.
- Legacy page-shell CSS removed (page-col, sa-view-*, sched-banner,
  tmpl-card family, sa-back, new-action).
- e2e: inbox routing test rewritten against the Configure tab;
  sidebar-account's Unrouted-location test follows the move.

tsc + 28 unit + 106 e2e green.
