+++ platform/surfaces/gui/src/components/InboxView.tsx
@@ -5,16 +5,17 @@ import {
   getInboxRouting,
   getPersonas,
   getRecentChannels,
+  getUnrouted,
   resolveInboxItem,
-  setInboxBinding,
   type InboxItem,
   type Persona,
   type RecentChannel,
 } from "../api";
 import { Icon } from "./Icon";
 import { InboxItemCard } from "./InboxItemCard";
+import { InboxConfigure } from "./InboxConfigure";
+import { PanelHead } from "./IntegrationsView";
 import { shortPersonaName } from "../personaScope";
-import { ChannelPicker } from "./SubscriptionsChip";
 
 const ICON_FOR: Record<string, "diamond" | "chat" | "code"> = {
   cowork: "diamond",
@@ -34,27 +35,41 @@ const CHIP = (active: boolean) =>
     ? "border-accent text-accent bg-accentSoft"
     : "border-line text-muted hover:border-lineStrong");
 
+// Page-level tabs (§28): underline style, one visual level ABOVE the filter chips.
+const TAB = (active: boolean) =>
+  "pb-2 -mb-px text-[13px] border-b-2 flex items-center gap-1.5 " +
+  (active
+    ? "text-ink font-medium border-accent"
+    : "text-muted border-transparent hover:text-ink");
+
 // The Inbox: pending approvals / questions / notifications from across sessions, including
 // unattended ones. Resolving here releases any agent suspended on the item. Each item links back
 // to its originating session so you can see the context before answering. Items whose session
 // was deleted are closed server-side (an orphaned prompt can never be answered), so everything
 // listed here is actionable. Filters: by kind and by persona (owner ask, 2026-07-03).
+// Two page tabs (§28): Pending (the queue) and Configure (the former Connectors ▸ Messaging
+// routing page — mirror channel, DM route, subscriptions, Unrouted). Pending's routing status
+// is read-only and links to Configure; the old inline editor was the mirror setting's SECOND
+// editor and is gone.
 export function InboxView({
   onOpenSession,
 }: {
   onOpenSession: (sessionId: string, workspace: string, agent: string) => void;
 }) {
+  const [tab, setTab] = useState<"pending" | "configure">("pending");
   const [items, setItems] = useState<InboxItem[]>([]);
   const [personas, setPersonas] = useState<Persona[] | null>(null);
   const [routing, setRouting] = useState<string | null>(null); // e.g. "slack:C0123" or null
   const [slackConnected, setSlackConnected] = useState(false);
-  const [configOpen, setConfigOpen] = useState(false);
   const [recent, setRecent] = useState<RecentChannel[]>([]);
-  const [draft, setDraft] = useState("");
+  const [unroutedCount, setUnroutedCount] = useState(0);
   const [kind, setKind] = useState<string>("all");
   const [personaFilter, setPersonaFilter] = useState<string>("all");
 
-  const load = () => getInbox(undefined, "pending").then(setItems).catch(() => {});
+  const load = () => {
+    getInbox(undefined, "pending").then(setItems).catch(() => {});
+    getUnrouted().then((u) => setUnroutedCount(u.length)).catch(() => setUnroutedCount(0));
+  };
   const loadRouting = () =>
     getInboxRouting()
       .then((bindings) => {
@@ -70,25 +85,12 @@ export function InboxView({
       .then((cs) => setSlackConnected(!!cs.find((c) => c.name === "slack" && c.connected)))
       .catch(() => {});
     getRecentChannels().then(setRecent).catch(() => setRecent([]));
-    const t = setInterval(load, 4000);
+    const t = setInterval(() => {
+      load();
+      loadRouting(); // edits happen on the Configure tab; keep Pending's status line honest
+    }, 4000);
     return () => clearInterval(t);
   }, []);
-
-  const saveRoute = async () => {
-    const raw = draft.trim();
-    if (!raw) return;
-    // Slack-only for now (Teams / Telegram / WhatsApp later): a bare id is a slack channel.
-    const [platform, id] = raw.includes(":") ? raw.split(":", 2) : ["slack", raw];
-    await setInboxBinding("default", platform, id);
-    setDraft("");
-    setConfigOpen(false);
-    loadRouting();
-  };
-  const clearRoute = async () => {
-    await setInboxBinding("default", null, "");
-    setConfigOpen(false);
-    loadRouting();
-  };
 
   const resolve = async (id: string, resolution: string) => {
     await resolveInboxItem(id, resolution);
@@ -135,98 +137,124 @@ export function InboxView({
     );
   };
 
+  const routingName = routing ? recent.find((c) => c.channel === routing)?.name : undefined;
+  const routingLabel = routingName ? `#${routingName}` : routing;
+
   return (
-    <div className="main-scroll">
-      <div className="page-col">
-        <div className="sa-view-head">
-          <div className="sa-view-title">Inbox</div>
-          <div className="sa-view-sub dim">
-            Approvals, questions, and notifications from your agents — including sessions running
-            unattended.
+    <main className="flex-1 min-w-0 flex bg-paper">
+      <div className="flex-1 min-w-0 overflow-y-auto hairline-scroll">
+        <div className="max-w-4xl mx-auto px-7 py-6">
+          <PanelHead
+            title="Inbox"
+            sub="Approvals, questions, and notifications from your coworkers — including sessions running unattended."
+          />
+
+          <div className="flex gap-5 border-b border-line mb-4">
+            <button
+              className={TAB(tab === "pending")}
+              data-testid="inbox-tab-pending"
+              onClick={() => {
+                setTab("pending");
+                // Configure-tab edits change the mirror target — re-read so the status line
+                // is honest the moment the user lands back on Pending, not a poll later.
+                loadRouting();
+                load();
+              }}
+            >
+              Pending
+              {items.length > 0 && (
+                <span className="text-[11px] px-1.5 rounded-full bg-accentSoft text-accent leading-4">
+                  {items.length}
+                </span>
+              )}
+            </button>
+            <button
+              className={TAB(tab === "configure")}
+              data-testid="inbox-tab-configure"
+              onClick={() => setTab("configure")}
+            >
+              Configure
+              {unroutedCount > 0 && (
+                <span className="text-[11px] px-1.5 rounded-full bg-warnSoft text-warnInk leading-4">
+                  ⚠ {unroutedCount}
+                </span>
+              )}
+            </button>
           </div>
-        </div>
-        <div className="text-[12px] text-faint -mt-2 mb-4" data-testid="inbox-routing">
-          {routing ? (
-            <span className="inline-flex items-center gap-2 flex-wrap">
-              <span>
-                Also delivered to <span className="text-muted">{routing}</span> — replies there
-                resolve items here.
-              </span>
-              <button className="text-accent hover:underline" onClick={() => setConfigOpen((v) => !v)}>
-                Change
-              </button>
-              <button className="text-danger/80 hover:text-danger" onClick={clearRoute}>
-                Stop
-              </button>
-            </span>
-          ) : slackConnected ? (
-            <span className="inline-flex items-center gap-2 flex-wrap">
-              <span>Delivered here only.</span>
-              <button
-                className="text-accent hover:underline"
-                data-testid="inbox-route-configure"
-                onClick={() => setConfigOpen((v) => !v)}
-              >
-                Also send to a Slack channel →
-              </button>
-            </span>
+
+          {tab === "configure" ? (
+            <InboxConfigure />
           ) : (
             <>
-              Delivered here only. Connect Slack (Connectors page) to also get these in a
-              channel — more platforms later.
-            </>
-          )}
-          {configOpen && (
-            <span className="flex items-center gap-2 mt-2">
-              <ChannelPicker value={draft} onChange={setDraft} recent={recent} onSubmit={saveRoute} />
-              <button
-                className="text-[12px] px-2.5 py-1 rounded-md bg-accent text-white disabled:opacity-50"
-                disabled={!draft.trim()}
-                onClick={saveRoute}
-              >
-                Set
-              </button>
-            </span>
-          )}
-        </div>
-
-        <div className="flex items-center gap-2 flex-wrap mb-4" data-testid="inbox-filters">
-          {KIND_TABS.map((t) => (
-            <button key={t.key} className={CHIP(kind === t.key)} onClick={() => setKind(t.key)}>
-              {t.label}
-            </button>
-          ))}
-          {personasWithItems.length > 1 && (
-            <>
-              <span className="w-px h-4 bg-line mx-1" />
-              <button className={CHIP(personaFilter === "all")} onClick={() => setPersonaFilter("all")}>
-                All coworkers
-              </button>
-              {personasWithItems.map((p) => (
+              <div className="text-[12px] text-faint -mt-1 mb-4" data-testid="inbox-routing">
+                {routing ? (
+                  <span>
+                    Also delivered to{" "}
+                    <span className="text-muted" title={routing}>
+                      {routingLabel}
+                    </span>{" "}
+                    — replies there resolve items here.{" "}
+                  </span>
+                ) : slackConnected ? (
+                  <span>Delivered here only. </span>
+                ) : (
+                  <span>
+                    Delivered here only. Connect Slack (Connectors page) to also get these in a
+                    channel — more platforms later.{" "}
+                  </span>
+                )}
                 <button
-                  key={p.id}
-                  className={CHIP(personaFilter === p.id)}
-                  onClick={() => setPersonaFilter(p.id)}
+                  className="text-accent hover:underline"
+                  data-testid="inbox-route-configure"
+                  onClick={() => setTab("configure")}
                 >
-                  {p.label}
+                  Configure ›
                 </button>
-              ))}
+              </div>
+
+              <div className="flex items-center gap-2 flex-wrap mb-4" data-testid="inbox-filters">
+                {KIND_TABS.map((t) => (
+                  <button key={t.key} className={CHIP(kind === t.key)} onClick={() => setKind(t.key)}>
+                    {t.label}
+                  </button>
+                ))}
+                {personasWithItems.length > 1 && (
+                  <>
+                    <span className="w-px h-4 bg-line mx-1" />
+                    <button
+                      className={CHIP(personaFilter === "all")}
+                      onClick={() => setPersonaFilter("all")}
+                    >
+                      All coworkers
+                    </button>
+                    {personasWithItems.map((p) => (
+                      <button
+                        key={p.id}
+                        className={CHIP(personaFilter === p.id)}
+                        onClick={() => setPersonaFilter(p.id)}
+                      >
+                        {p.label}
+                      </button>
+                    ))}
+                  </>
+                )}
+              </div>
+
+              {visible.length === 0 ? (
+                <div className="manage-empty">
+                  {items.length === 0 ? "Nothing pending." : "Nothing pending for this filter."}
+                </div>
+              ) : null}
+
+              <div className="space-y-4">
+                {visible.map((it) => (
+                  <InboxItemCard key={it.id} item={it} onResolve={resolve} chip={sessionChip(it)} />
+                ))}
+              </div>
             </>
           )}
-        </div>
-
-        {visible.length === 0 ? (
-          <div className="manage-empty">
-            {items.length === 0 ? "Nothing pending." : "Nothing pending for this filter."}
-          </div>
-        ) : null}
-
-        <div className="space-y-4">
-          {visible.map((it) => (
-            <InboxItemCard key={it.id} item={it} onResolve={resolve} chip={sessionChip(it)} />
-          ))}
         </div>
       </div>
-    </div>
+    </main>
   );
 }
